+++ src/helpers/ScreenShotService/BMP/TinyBMPEncoder.cpp
