+++ src/helpers/ScreenShotService/JPG/TinyJPEGEncoder.cpp
