+++ src/helpers/ScreenShotService/JPG/TinyJPEGEncoder.h
@@ -56,7 +56,7 @@
 #ifndef __TINY_JPEG_ENCODER__
 #define __TINY_JPEG_ENCODER__
 
-#include "../M5Display.h"
+#include "../../../M5Display.h"
 // C std lib
 #include <assert.h>
 #include <inttypes.h>
