+++ src/helpers/ScreenShotService/PNG/FatPNGEncoder.cpp
