+++ src/helpers/ScreenShotService/ScreenShot.cpp
